Added BTI POC example
This test contains the code for running a BTI attack on a victim
process. The attacker runs and poison the BTB and the victim is hijacked
towards code that is not supposed to be speculated where we conveniently
placed markers which will indicate when the attack has been run
successfully.