This commit adds the ability to run speculator and its tests as
monitor-only. This means that no configuration file is read, no output
files are generated and there is no setting/reading of counters.
On the snippet side, the macros start/stop/reset counters are skipped
since the uid running is not root (0).

Not all the cases have been tested yet. Further changes might be needed.